Added support for micrometer in hermes-consumers

+++ hermes-metrics/src/main/java/pl/allegro/tech/hermes/metrics/HermesHistogram.java
@@ -15,7 +15,7 @@ public class HermesHistogram {
         return new HermesHistogram(micrometerHistogram, graphiteHistogram);
     }
 
-    public void record(int value) {
+    public void record(long value) {
         micrometerHistogram.record(value);
         graphiteHistogram.update(value);
     }
